Vastly improved responsiveness of play/pause.
This was done by simply only using synchronized for the wait/notify
communication between threads for animation restarts.

State (in native variables) that is manipulated between threads is
marked volatile so that reads/writes are atomic.  (Note, this is actually only
necessary for long and double, but done for other the booleans anyway.)

Of interest is that this was not an issue back in 2002, animation pause/restart
was responsive back then, but probably relied on undocumented behaviour of the JVM.